+++ mod/settings.php
@@ -24,7 +24,7 @@ use Friendica\Model\Contact;
 use Friendica\Model\GContact;
 use Friendica\Model\Group;
 use Friendica\Model\User;
-use Friendica\Module\Login;
+use Friendica\Module\Security\Login;
 use Friendica\Protocol\Email;
 use Friendica\Util\ACLFormatter;
 use Friendica\Util\Network;
